requirements.txt
# GenAI Frameworks
langchain==0.3.4
langchain-groq==0.2.1
langchain-community==0.3.3
langchain-huggingface==0.1.2

python-dotenv==1.0.1
sse-starlette==2.1.3

#cloud db
langchain-pinecone==0.2.5

# Backend - RestAPI
uvicorn==0.32.1
fastapi==0.115.5
langserve==0.3.0
python-multipart==0.0.17
# joblib==1.4.2
httpx==0.27.2
pypdf
streamlit==1.40.2